refactor script engine eval codes
