+++ hyperdbg/hprdbghv/hprdbghv.vcxproj.filters
@@ -159,6 +159,9 @@
     <Filter Include="header\debugger\script-engine">
       <UniqueIdentifier>{5b6684f0-db1b-4920-987c-d5dfd8cdc0c1}</UniqueIdentifier>
     </Filter>
+    <Filter Include="code\script-eval">
+      <UniqueIdentifier>{59005a75-5137-42a4-867a-ccf06fae2f9f}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="code\common\Common.c">
@@ -326,8 +329,20 @@
     <ClCompile Include="pch.c">
       <Filter>code</Filter>
     </ClCompile>
-    <ClCompile Include="..\script-eval\ScriptEngineEval.c">
-      <Filter>code\debugger\script-engine</Filter>
+    <ClCompile Include="..\script-eval\code\ScriptEngineEval.c">
+      <Filter>code\script-eval</Filter>
+    </ClCompile>
+    <ClCompile Include="..\script-eval\code\PseudoRegisters.c">
+      <Filter>code\script-eval</Filter>
+    </ClCompile>
+    <ClCompile Include="..\script-eval\code\Keywords.c">
+      <Filter>code\script-eval</Filter>
+    </ClCompile>
+    <ClCompile Include="..\script-eval\code\Functions.c">
+      <Filter>code\script-eval</Filter>
+    </ClCompile>
+    <ClCompile Include="..\script-eval\code\Regs.c">
+      <Filter>code\script-eval</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
