+++ hyperdbg/hprdbghv/pch.c
@@ -1,1 +1,13 @@
+/**
+ * @file pch.c
+ * @author Sina Karvandi (sina@hyperdbg.org)
+ * @brief Pre-compiled headers
+ * @details
+ *
+ * @version 0.1
+ * @date 2022-06-29
+ *
+ * @copyright This project is released under the GNU Public License v3.
+ *
+ */
 #include "pch.h"
