+++ hyperdbg/hyperdbg.sln
@@ -65,10 +65,21 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Headers", "Headers", "{D67D
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "script-eval", "script-eval", "{2CC2975A-9B33-400D-840F-9A71BB312928}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "header", "header", "{4C47AB91-DF41-4741-80AF-FD87C32A961E}"
 	ProjectSection(SolutionItems) = preProject
-		script-eval\ScriptEngineCommonDefinitions.h = script-eval\ScriptEngineCommonDefinitions.h
-		script-eval\ScriptEngineEval.c = script-eval\ScriptEngineEval.c
-		script-eval\ScriptEngineHeader.h = script-eval\ScriptEngineHeader.h
+		script-eval\header\ScriptEngineCommonDefinitions.h = script-eval\header\ScriptEngineCommonDefinitions.h
+		script-eval\header\ScriptEngineHeader.h = script-eval\header\ScriptEngineHeader.h
+		script-eval\header\ScriptEngineInternalHeader.h = script-eval\header\ScriptEngineInternalHeader.h
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "code", "code", "{952986BC-7B3D-4F50-BC69-4D11FA52D9C5}"
+	ProjectSection(SolutionItems) = preProject
+		script-eval\code\Functions.c = script-eval\code\Functions.c
+		script-eval\code\Keywords.c = script-eval\code\Keywords.c
+		script-eval\code\PseudoRegisters.c = script-eval\code\PseudoRegisters.c
+		script-eval\code\Regs.c = script-eval\code\Regs.c
+		script-eval\code\ScriptEngineEval.c = script-eval\code\ScriptEngineEval.c
 	EndProjectSection
 EndProject
 Global
@@ -116,6 +127,8 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{7EA9E1E8-4283-4CEF-9629-C5C98CD706C7} = {1997E4B6-FE49-4358-ADF1-CDDB5739508E}
 		{D67D603F-4F99-4C0B-94BE-F4A2C6CC22B5} = {7EA9E1E8-4283-4CEF-9629-C5C98CD706C7}
+		{4C47AB91-DF41-4741-80AF-FD87C32A961E} = {2CC2975A-9B33-400D-840F-9A71BB312928}
+		{952986BC-7B3D-4F50-BC69-4D11FA52D9C5} = {2CC2975A-9B33-400D-840F-9A71BB312928}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1444BEC7-11CE-4CA6-B77C-5F98AC9BFAEB}
