+++ hyperdbg/script-eval/header/ScriptEngineCommonDefinitions.h
