+++ hyperdbg/script-eval/header/ScriptEngineHeader.h
